+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Released versions should use the Git tag as the section title, for example
 ### Added
 
 - IP anonymization prefix lengths are now configurable.
+- Log a warning if a path could not be fully unquoted.
 
 ### Changed